+++ CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [2.5.4] 2025-03-23
 
+### Fixed
+- backend deployment failure due to existing AppPlan with same name in different reqion.
+- backend deployment failure due to existing event grid system topic.
 
 ## [2.5.3] 2025-02-12